pyproject.toml
[project]
name = "shape-recognition-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pandas-stubs~=2.3.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "telebot>=0.0.5",
    "tensorflow>=2.20.0",
]
